Fix weird usage of ElectrolysisSetup::current_source
This was clearly intended to store a reference to an object, but in
practice, it was usually used as a `bool` while sometimes being set to
`null`. In my opinion, the best solution is to just name it
`has_current_source` and lean all the way into the `bool` approach. If
we eventually need to store a reference to an actual object, it can
easily be changed.