Make *Content functions consistent
`CheckContents` in particular would sometimes return an `Array[bool]`
and sometimes a `bool`. So I made it so all instances of `CheckContents`
would return `Array[bool]`. Any cases that looked like

```
if o.CheckContents(g)
```

Are instead written as

```
if o.CheckContents(g).front()
```

This works because the `front` function will return `null` if the first
element doesn't exist, which also evaluates to false.