Handle if expression type issues
For some reason, when used as one of the alternatives of an `if`
expression, an empty array is treated as just being of type `Array`
even if it's used to initialize a typed array. This causes a runtime
error instead of a static type error. So they have to be replaced here
with regular `if` statements.