Fix weird micropipette plunger behavior
I think previously, the `Slider.value_changed` signal would only be
emitted when the value changed to a *different* value, so that behavior
could be relied on. Now, this signal is emitted constantly while the
slider is being dragged. So when the plunger is at the top, we can't
assume that it *changed* to the top from a lower position; we have to
specifically check to make sure it has plunged a nonzero amount first.