Fix AttachmentPoints with unmodified positions not working
`AttachmentInteractableArea::_find_attachment_offset` returns an
optional `Vector2`, but since that's not available in gdscript, I
instead made it return a `Variant`. In this case, I had the return value
of the function stored in a variable `offset` and checked whether it
existed with

```
if offset:
	# Do something
```

However, even if the offset is not null but is still `Vector2.ZERO`,
then it will still be contextually evaluated to false, since
`Vector2.ZERO` is "false". So it would refuse to properly attach if the
`AttachmentPoint` were not moved, which is clearly incorrect.