Interacting with area directly
Instead of using get_overlapping_bodies, the scooping system interacts with the `ScoopInteractableArea` directly. Also, the scoopula system has been updated to be left click to pick up 0.1 mL of a substance and right click to dispense 0.1 mL of a substance. Lastly, the scale logic has been updated such that it only weighs objects with the `WeighInteractableArea` so that the scale will disregard the scoopula when dispensing product into the weightboat while it's on the scale.