Revert partially merged changes from dev
These changes to `project.godot` include some changes from later
versions of the `dev` branch, but not all of the relevant changes seem
to have been added, which is causing errors (e.g., the `LabTime`
singleton was added to `project.godot`, but its script
`lab_time_singleton.gd` was not). I'm removing this to make it a bit
easier to merge this branch into `dev`.