Make every object weighable
`WeighInteractableArea`, as far as I understand, was not really being
used for interaction, but was mainly being used as a sort of "tag" to
mark objects as weighable. My assumption is that this was to prevent the
scoopula from being weighed while it's being held near the weigh boat on
top of the scale. If we wanted to go down the route of manually marking
things as weighable, then I think a more appropriate method would be to
use groups (i.e., the weigh boat would be put in the `weighable` group,
and the scale would check for `body.is_in_group("weighable")`).

However, I think a more straightforward solution is to just only weigh
objects if they are affected by gravity (the held object is not affected
by gravity). In a more recent version of the `dev` branch, this can be
checked with `body.physics_mode == LabBody.PhysicsMode.FREE`. But since
this branch isn't caught up with that version of `dev`, a decent
alternative is to just check that `body.freeze` is false, since
free-falling lab bodies have `freeze` set to false. It can be changed to
use the `physics_mode` check when this branch is merged into `dev`.