Fix ScoopulaUseComponent.start_use parameters
The parameter names of a function overriding another function do not
have to match the names of the parameters of the original function; only
the parameter types and number of parameters have to match. In this
case, the base class `UseComponent` has a function

```
func start_use(_area: InteractableArea, _kind: InteractInfo.Kind) -> void: pass
```

Just as a GDScript convention, names of unused variables and parameters
start with an underscore (this will prevent the "unused variable"
warning" for that variable). So because those parameters aren't used in
the base class, the parameters in the base class start with an
underscore.

However, because the parameter names don't matter when overriding, any
parameter that's actually *used* in the override can be renamed to no
longer include the underscore. In this case, both `_area` and `_kind`
are used, so they should be renamed to `area` and `kind`, respectively.
So the overridden function can look like:

```
func start_use(area: InteractableArea, kind: InteractInfo.Kind) -> void:
	# `ScoopulaUseComponent` use code.
```

The class-wide `area` variable is not needed for this.