Use names of enumerators from InteractInfo.Kind instead of numbers
The numeric values of the enumerators (e.g., the value 0 for
`InteractInfo.Kind.PRIMARY`) are not guaranteed to remain the same if
new new kinds are added, so this code could lead to some errors in the
future that are hard to track. I think it's also more clear to use the
names.