Replace ContainerComponent.get_volume with an exported container_volume variable
This value needs to be an exported variable so that different containers
can have their volume set to different quantities in the editor rather
than having a fixed 100 mL volume for every container.

There is already an exported `ContainerComponent.max_volume` variable in
the `dev` branch which serves the same purpose, but I realized as I was
making this change that `max_volume` is not a good name for this
variable, since it implies that the container itself will prevent the
volume from exceeding it (which it doesn't do). I ended up going with
`container_volume`, which is also kind of bad name in my opinion, but
I'm not sure I can come up with anything better at the moment. It can
probably be changed to something better in the merge.