Fix the scale showing incorrect values for multiple objects or when one is removed after taring
Since we're already calculating the total weight and setting the label
text every tick, it's not necessary to add special handling for when an
object is removed. On a real scale, if you place an object on it, tare
it, and then remove the object, then the scale will show a negative
weight. Setting `current_weight` and `tare_weight` to zero when removing
an object causes it to show zero instead of the correct negative value.