Include the weight of every direct ContainerComponent child when weighing on the scale
Some objects, like the gel mold, can have multiple `ContainerComponent`s
that should all probably be accounted for when taking their weight.
Doing it this way also ensures that we don't have to worry about
ordering `ContainerComponent`s in an object to make sure that the first
one is the "correct" one.

This *still* doesn't properly handle more complex weighing situations,
like when an object is attached to the object being weighed. I'm not
sure how to best handle this here, but if we had interfaces, then I
would probably make some sort of `Weighable` interface that the
`AttachmentInteractableArea`s could then implement.