Add interface of gel band functions to the imager class
The gel state will all be data from a Gel class that will be passed to the imager along with the state of the UV (whether it was turned on before imaging or not). If the UV_state is false, a gel blank sprite will be shown upon the user pulling the gel mold out of the imager. Otherwise, the imager will match the gel state to the appropriate gel band sprite. When the user removes the gel from the imager, the original gel sprite will be replaced by the updated sprite with gel bands in it.