Fix Interaction inputs taking priority over menu inputs
Since viewports automatically mark inputs as handled, any input would be
absorbed by `$%MainViewport`, so they would never be sent to the
`_unhandled_input` function of the `Interaction` singleton. To fix this,
I changed it to instead use the `_input` function. But this creates a
different problem: `_input` happens *before* `_gui_input`, so this would
take priority over the menu UI, which prevented it from working
properly.

The solution to this is to move the actual `_unhandled_input` function
out of the `Interaction` singleton and into a node under the main scene
viewport.