Fix SubsceneCamera not rendering the correct location
The engine currently has a bug where it uses the size of the parent
viewport instead of `custom_viewport` when calculating the position to
render. When using the centered anchor mode, this causes the viewport to
be offset by the incorrect amount. However, the size is not used when
calculating the position in top left mode, so setting the anchor mode to
`ANCHOR_MODE_FIXED_TOP_LEFT` does the correct thing.