Add a virtual cursor active when not paused
When the simulation is not paused, the actual cursor is locked, and a
"virtual" cursor is moved around in its stead. This prevents it from
jumping around when the game is unpaused, and ensures that it retains
its world space position even when zooming in (preventing weird behavior
where a held object will move around when zoomed in).