Make `InteractInfo` into an exportable resource
The intent is to allow certain interactables (like
`SelectableComponent`s) to just directly export the `InteractInfo` they
will use instead of having to export all of the parameters separately.

The `_init` function also needed to be given default values for all of
its parameters, since exported resources need to be constructible with
just `_init()` (no arguments) to be used in scenes.