Greatly simplify substances
The substances can be simplified a lot by just moving most of the logic
into each substance on a case-by-case basis rather than trying to have a
general system for solubility and stuff.

So I've removed `SolutionSubstance` and simplified `BasicSubstance` by
removing the whole `BasicSubstanceData` thing and renaming it to
`GenericSubstance`. I have also removed the mixing and temperature stuff
from `ContainerComponent`.