Clean up camera zoom logic
This adds a "camera focus owner" system where a node can tell `Main`
that it is currently zooming in. When there is an active focus owner,
`Main` will provide a keybinding to zoom out and will stop allowing the
user to move to the left and right workspace. There is also a new
`Main.camera_focus_owner_changed` signal that nodes can use to detect
when a different node has taken camera focus or when it's zoomed out to
the current workspace.