Fix some resources not fully reloading when restarting a module
This was primarily noticeable with substances, which would carry over
even after restarting the module. For example, if DNA was transferred
from one microcentrifuge tube to another, then the first tube would
remain empty and the new one would still have the transferred content.

Modules have been handled by first loading them into a `PackedScene`,
then calling `instantiate` on that packed scene when starting or
restarting the module. However, it turns out that a `PackedScene` comes
with only a single set of shared resources, so every instance would be
using the same instances of each resource.

This was fixed by replacing the `PackedScene` in each module with a path
to the module's scene, and just calling `load` every time we want to
load the module. This makes a completely fresh instance of the module
with new resources and everything.