Fix tertiary interaction being started multiple times before stopping
This was especially weird, because it happened on Windows but not on
MacOS. The problem was that we were counting key echoes (i.e., repeated
presses of keys when they're held down) as the key being "just pressed
down", and relying on it only being pressed down once before being
released. The tertiary interaction uses shift by default, which Windows
generates echoes for, but MacOS does not.