Improve interaction targeting behavior
We had an issue where we could pick up an object, and it would stay
highlighted because the "put down" interaction was not considered
different from the "pick up" interaction, so it would not call
`stop_targeting` for the pick up interaction after picking up. This has
been fixed by adding derived types for `InteractState` that are able to
determine whether a particular interaction is equivalent to them. In
this case, the "pick up" interaction falls under
`ComponentInteractState`, and the "put down" falls under
`UseInteractState`, which makes them different, resulting in a retarget.