Actually fix the scaled button prompts for real this time
Changing the size of a control node does not cause it to automatically
update its position to match its anchors. Ideally, if a control is
supposed to automatically be repositioned, then it should be put into a
container. However, containers do not allow you to scale their children;
their children will always be at a scale of 1.

My solution to this is to make a new `ScaleContainer` container that
automatically scales its children and shrinks to fit around them. Since
the container itself is not scaled, it can safely be used in a
container. I also moved the prompts into a `MarginContainer` and changed
their container layout flags to position them correctly rather than
using their anchors directly.