Fix web export not working properly on MacOS
It seems that the `.pck` file wasn't being found because it had capital
letters in it, and MacOS is a case-insensitive filesystem.

I'm not 100% sure this was the cause, but it was fixed by using only
lowercase letters in the name of the exported file, so that's my best
guess. I also don't know whether the same issue occurs on Windows or
Linux.