Add Power Supply Error and Imager door
If there's no TAE Buffer solution when trying to run a current through the gel, an error while appear on the power supply's screen. Also, users can now load and remove gel from the imager by clicking on the door and opening the imager in a similar fashion as the microwave.