Fix errors from `SubstanceDisplayPolygon` while adding new substances
Since elements in the substance array default to null in the editor
before you add them, `SubstanceDisplayPolygon` was causing errors when
trying to call functions on the partially added substances in the
editor. This was fixed by ensuring that it only shows "safe" substances
in the editor (i.e., substances that are not null and use tool scripts).